requirements.txt
flask  
pdfplumber  
python-docx  
spacy  
fpdf 
